+++ @types/react-i18next.d.ts
@@ -1,6 +1,6 @@
 import "react-i18next";
 
-import { defaultNS, ResourcesType } from "../src/model/translation";
+import { defaultNS, ResourcesType } from "../src/model/i18next";
 
 declare module "react-i18next" {
   interface CustomTypeOptions {
